Module Av.Format

FFmpeg's input and output formats: demuxers, muxers and devices. A format value is a handle on a static FFmpeg object and needs no release.

val get_input_name : (Avutil.input, _) Avutil.format -> string

The name of an input format. It may be a comma-separated list of aliases.

val get_input_long_name : (Avutil.input, _) Avutil.format -> string

The descriptive name of an input format.

val get_output_name : (Avutil.output, _) Avutil.format -> string

The name of an output format. It may be a comma-separated list of aliases.

val get_output_long_name : (Avutil.output, _) Avutil.format -> string

The descriptive name of an output format.

val find_input_format : string -> (Avutil.input, 'a) Avutil.format option

find_input_format name is FFmpeg's input format of that short name, or None when there is none. The media parameter of the result is chosen by the caller and is not checked.

val guess_output_format : ?short_name:string -> ?filename:string -> ?mime:string -> unit -> (Avutil.output, 'a) Avutil.format option

FFmpeg's guess of an output format. The media parameter of the result is chosen by the caller and is not checked.

parameter short_name
a format short name, such as "mp4".

parameter filename
a file name.

parameter mime
a MIME type.

returns
None when FFmpeg guesses nothing. An omitted or empty argument does not take part in the guess.

val get_audio_codec_id : (Avutil.output, Avutil.audio) Avutil.format -> Avcodec.Audio.id

The default audio codec of the format; `None when it has none.

val get_video_codec_id : (Avutil.output, Avutil.video) Avutil.format -> Avcodec.Video.id

The default video codec of the format; `None when it has none.

val get_subtitle_codec_id : (Avutil.output, Avutil.subtitle) Avutil.format -> Avcodec.Subtitle.id

The default subtitle codec of the format; `None when it has none.