Module Av.Format

val get_input_name : (Avutil.input, _) Avutil.format -> string

Return the name of the input format

val get_input_long_name : (Avutil.input, _) Avutil.format -> string

Return the long name of the input format

val find_input_format : string -> (Avutil.input, 'a) Avutil.format option

Guess input format based on its short name.

val get_output_name : (Avutil.output, _) Avutil.format -> string

Return the name of the output format

val get_output_long_name : (Avutil.output, _) Avutil.format -> string

Return the long name of the output format

val guess_output_format : ?short_name:string -> ?filename:string -> ?mime:string -> unit -> (Avutil.output, 'a) Avutil.format option

Guess output format based on the passed arguments.

val get_audio_codec_id : (Avutil.output, Avutil.audio) Avutil.format -> Avcodec.Audio.id

Return the audio codec id of the output audio format

val get_video_codec_id : (Avutil.output, Avutil.video) Avutil.format -> Avcodec.Video.id

Return the video codec id of the output video format

val get_subtitle_codec_id : (Avutil.output, Avutil.subtitle) Avutil.format -> Avcodec.Subtitle.id

Return the subtitle codec id of the output subtitle format